Do pointer math only with "ptrdiff_t".

+++ oncrpc/win32/librpc/xdr_mem.c
@@ -176,7 +176,7 @@ xdrmem_getpos(xdrs)
 	register XDR *xdrs;
 {
 
-	return ((u_int)xdrs->x_private - (u_int)xdrs->x_base);
+	return ((ptrdiff_t)xdrs->x_private - (ptrdiff_t)xdrs->x_base);
 }
 
 static bool_t
@@ -187,10 +187,10 @@ xdrmem_setpos(xdrs, pos)
 	register caddr_t newaddr = xdrs->x_base + pos;
 	register caddr_t lastaddr = xdrs->x_private + xdrs->x_handy;
 
-	if ((long)newaddr > (long)lastaddr)
+	if ((ptrdiff_t)newaddr > (ptrdiff_t)lastaddr)
 		return (FALSE);
 	xdrs->x_private = newaddr;
-	xdrs->x_handy = (int)lastaddr - (int)newaddr;
+	xdrs->x_handy = (ptrdiff_t)lastaddr - (ptrdiff_t)newaddr;
 	return (TRUE);
 }
